Separated test for SigmaJs and xml_parser, and rework function for loading json
