+++ tests/any_test_help.py
@@ -67,13 +67,12 @@ def any_test_transformation_tree_to_Json_for_SigmaJs(
 
     oval_tree = graph.oval_graph.build_nodes_form_xml(src, rule_id)
 
-    if oval_tree.node_id == rule_id:
-        out_data = oval_tree.to_sigma_dict(0, 0)
-        print(json.dumps(out_data))
-        for i in range(len(out_data['nodes'])):
-            assert out_data['nodes'][i]['label'] == test_data['nodes'][i]['label']
-            assert out_data['nodes'][i]['text'] == test_data['nodes'][i]['text']
-            assert out_data['nodes'][i]['url'] == test_data['nodes'][i]['url']
+    assert oval_tree.node_id == rule_id
+    out_data = oval_tree.to_sigma_dict(0, 0)
+    for i in range(len(out_data['nodes'])):
+        assert out_data['nodes'][i]['label'] == test_data['nodes'][i]['label']
+        assert out_data['nodes'][i]['text'] == test_data['nodes'][i]['text']
+        assert out_data['nodes'][i]['url'] == test_data['nodes'][i]['url']
 
 
 def any_test_tree_to_dict_of_tree(tree, dict_of_tree):
